AVM support

+++ aion-api/src/main/java/com/aion/dashboard/services/BlockService.java
@@ -173,7 +173,7 @@ public class BlockService {
 							txn.getTransactionHash() + "," +
 							txn.getFromAddr() + "," +
 							txn.getToAddr() + "," +
-							Utility.toAion(txn.getValue()).doubleValue() + "," +
+							Utility.toAion(txn.getValue()) + "," +
 							txn.getBlockTimestamp() + "," +
 							txn.getBlockNumber() +
 							"]";
@@ -211,7 +211,7 @@ public class BlockService {
 							txn.getTransactionHash() + "," +
 							txn.getFromAddr() + "," +
 							txn.getToAddr() + "," +
-							Utility.toAion(txn.getValue()).doubleValue() + "," +
+							Utility.toAion(txn.getValue()) + "," +
 							txn.getBlockTimestamp() + "," +
 							txn.getBlockNumber() +
 							"]";
